# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend Compro API Documentation",
        "description": "Dokumentasi API untuk Backend Company Profile Surya Kencana\n * ## Fitur Utama\n * **Authentication** - Login dan logout untuk admin/super admin\n * **Products** - Manajemen produk dengan upload gambar\n * **Vision & Mission** - Manajemen visi dan misi perusahaan\n * **Clients** - Manajemen data client/partner dengan upload logo\n * **Testimonials** - Manajemen testimoni pelanggan\n * **Contact** - Manajemen informasi kontak perusahaan\n * **Admin Management** - Manajemen akun admin (hanya super admin)\n * ## Cara Menggunakan\n * **Langkah 1:** Login melalui endpoint POST `/api/login`\n * **Langkah 2:** Copy token dari response JSON\n * **Langkah 3:** Klik tombol **Authorize** (gembok hijau di atas)\n * **Langkah 4:** Paste token dengan format `Bearer {your-token}`\n * **Langkah 5:** Klik Authorize dan Close\n * **Langkah 6:** Sekarang Anda bisa akses endpoint yang memerlukan autentikasi",
        "contact": {
            "name": "API Support",
            "email": "support@suryakencana.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Alternative Local Server"
        }
    ],
    "paths": {
...
        "/api/our-client": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Ambil semua data client",
                "description": "Mendapatkan daftar semua client/partner perusahaan beserta logo mereka",
                "operationId": "getClients",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "client_name": {
                                                "type": "string",
                                                "example": "PT ABC Technology"
                                            },
                                            "institution": {
                                                "type": "string",
                                                "example": "Technology Company"
                                            },
                                            "logo_path": {
                                                "type": "string",
                                                "example": "ourClients/abc123.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Tambah client baru",
                "description": "Membuat client baru dengan upload logo. Endpoint ini memerlukan autentikasi.",
                "operationId": "storeClient",
                "requestBody": {
                    "description": "Data client yang akan dibuat (multipart/form-data)",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "client_name",
                                    "institution",
                                    "logo_path"
                                ],
                                "properties": {
                                    "client_name": {
                                        "description": "Nama client/partner",
                                        "type": "string",
                                        "example": "PT ABC Technology"
                                    },
                                    "institution": {
                                        "description": "Nama institusi/jenis perusahaan",
                                        "type": "string",
                                        "example": "Technology Company"
                                    },
                                    "logo_path": {
                                        "description": "File logo client (jpeg, png, jpg, gif, svg, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Client created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "client_name": {
                                                    "type": "string",
                                                    "example": "PT ABC Technology"
                                                },
                                                "institution": {
                                                    "type": "string",
                                                    "example": "Technology Company"
                                                },
                                                "logo_path": {
                                                    "type": "string",
                                                    "example": "ourClients/abc123.jpg"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "client_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The client name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid atau tidak ada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Bearer yang didapat dari endpoint /api/login. Format: Bearer {your-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}